refactor(monitoreo): WebhookLog como ResourceBase
