+++ initialize.ts
@@ -61,7 +61,8 @@ export function initAPI(app: Express) {
 
     app.use('/api/modules/gestor-usuarios', require('./modules/gestor-usuarios').UsuariosRouter);
     app.use('/api/modules/gestor-usuarios', require('./modules/gestor-usuarios').PerfilesRouter);
-
+    app.use('/api/modules', require('./modules/webhooklog').WebhookLogRouter);
+    
     /**
      * Inicializa las rutas para adjuntar archivos
      */
