+++ modules/webhooklog/index.ts
@@ -1,0 +1,1 @@
+export { WebhookLogRouter } from './webhooklog.routes';
